requirements.txt
altair==5.3.0
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.8.2
branca==0.7.2
bs4==0.0.1
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.20.0
flake8==6.1.0
folium==0.17.0
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
haversine==2.8.1
idna==3.4
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy==1.4.1
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.2.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
prompt-toolkit==3.0.43
protobuf==5.27.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pycodestyle==2.11.0
pycparser==2.21
pydeck==0.9.1
pyflakes==3.1.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
pyzmq==26.0.2
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.11.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.3
streamlit==1.36.0
streamlit_folium==0.20.1
sweetviz==2.3.1
tenacity==8.4.1
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.10.3
types-python-dateutil==2.8.19.14
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.0.4
watchdog==4.0.1
wcwidth==0.2.13
wsproto==1.2.0
xyzservices==2024.6.0